# api/openapi.yaml
openapi: 3.1.0
info:
  title: scanX API
  version: 1.0.0-phase2
  description: |
    scanX REST API (spec §9). Authenticate with `Authorization: Bearer scanx_pat_…`
    (personal API token) or with the browser session cookie plus `X-CSRF-Token`.
    Resources of organizations you do not belong to are reported as 404.
  license:
    name: BUSL-1.1
    url: https://github.com/ininia/scanx/blob/main/LICENSE
servers:
  - url: /api/v1
security:
  - bearer: []
  - session: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Log in with e-mail and password (sets the session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Logged in; if mfa_required, POST /auth/mfa next
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /orgs/{org}/projects/{project}/scans:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/project" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Recent scans (newest first)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Scan" } }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [scans]
      summary: Queue a scan (default branch when omitted)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                branch: { type: string }
      responses:
        "202":
          description: Queued (or the already queued scan of this branch)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /orgs/{org}/scans/{scan}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/scan" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Scan with tool runs
          content:
            application/json:
              schema:
                type: object
                required: [scan, project, tools]
                properties:
                  scan: { $ref: "#/components/schemas/Scan" }
                  project: { type: string }
                  tools: { type: array, items: { type: object } }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/scans/{scan}/issues:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/scan" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Issues found by this scan
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Issue" } }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    session: { type: apiKey, in: cookie, name: __Host-scanx_session }
  parameters:
    org: { name: org, in: path, required: true, schema: { type: string } }
...
    project: { name: project, in: path, required: true, schema: { type: string } }
    scan: { name: scan, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            field: { type: string }
            reason: { type: string }
...
    LoginResult:
      type: object
      required: [mfa_required]
      properties:
        mfa_required: { type: boolean }
...
    Scan:
      type: object
      required: [id, project_id, trigger, branch, commit_sha, status, partial, new_issues, fixed_issues, summary, queued_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        trigger: { type: string, enum: [webhook, schedule, manual, api, ci_upload] }
        branch: { type: string }
        commit_sha: { type: string }
        commit_message: { type: string }
        status: { type: string, enum: [queued, cloning, scanning, reporting, completed, failed, canceled] }
        status_reason: { type: string }
        partial: { type: boolean }
        gate_result: { type: [string, "null"], enum: [pass, fail, warn, null] }
        score: { type: [integer, "null"] }
        new_issues: { type: integer }
        fixed_issues: { type: integer }
        summary: { type: object }
        queued_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
    Issue:
      type: object
      required: [id, fingerprint, category, severity, title, status]
      properties:
        id: { type: string, format: uuid }
        fingerprint: { type: string }
        category: { type: string }
        severity: { type: string, enum: [critical, high, medium, low, info] }
        title: { type: string }
        rule_id: { type: string }
        file: { type: string }
        start_line: { type: integer }
        cwe: { type: array, items: { type: string } }
        cve: { type: array, items: { type: string } }
        sources: { type: array, items: { type: string } }
        status: { type: string, enum: [open, fixed, false_positive, accepted_risk, wont_fix] }
        is_new: { type: boolean }
...