# internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type ApiToken struct {
	ID          uuid.UUID
	OrgID       uuid.UUID
	UserID      *uuid.UUID
	ProjectID   *uuid.UUID
	Name        string
	TokenPrefix string
	TokenHash   []byte
	Scopes      []string
	LastUsedAt  *time.Time
	ExpiresAt   *time.Time
	RevokedAt   *time.Time
	CreatedAt   time.Time
}

type AuditLog struct {
	ID         uuid.UUID
	OrgID      *uuid.UUID
	UserID     *uuid.UUID
	Action     string
	TargetType string
	TargetID   string
	Ip         string
	Metadata   json.RawMessage
	CreatedAt  time.Time
}

type InstanceSetting struct {
	Key       string
	Value     json.RawMessage
	UpdatedAt time.Time
}

type Invitation struct {
	ID         uuid.UUID
	OrgID      uuid.UUID
	Email      string
	Role       string
	TokenHash  []byte
	InvitedBy  *uuid.UUID
	ExpiresAt  time.Time
	AcceptedAt *time.Time
	CreatedAt  time.Time
}

type Issue struct {
	ID              uuid.UUID
	OrgID           uuid.UUID
	ProjectID       uuid.UUID
	Fingerprint     string
	Category        string
	Severity        string
	Title           string
	RuleID          string
	File            string
	StartLine       int32
	Cwe             []string
	Cve             []string
	Sources         []string
	Status          string
	StatusReason    string
	Detail          json.RawMessage
	FirstSeenScanID *uuid.UUID
	LastSeenScanID  *uuid.UUID
	FixedInScanID   *uuid.UUID
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type IssueEvent struct {
	ID        uuid.UUID
	OrgID     uuid.UUID
	IssueID   uuid.UUID
	UserID    *uuid.UUID
	Kind      string
	FromValue string
	ToValue   string
	Comment   string
	CreatedAt time.Time
}

type Job struct {
	ID          uuid.UUID
	OrgID       uuid.UUID
	Kind        string
	ProjectID   uuid.UUID
	ScanID      *uuid.UUID
	Status      string
	Result      json.RawMessage
	Error       string
	Attempts    int32
	LockedBy    string
	LockedUntil *time.Time
	CreatedAt   time.Time
	StartedAt   *time.Time
	FinishedAt  *time.Time
}

type Membership struct {
	OrgID     uuid.UUID
	UserID    uuid.UUID
	Role      string
	CreatedAt time.Time
}

type NotificationChannel struct {
	ID        uuid.UUID
	OrgID     uuid.UUID
	Kind      string
	Name      string
	TargetEnc []byte
	Nonce     []byte
	Events    []string
	CreatedAt time.Time
}

type Organization struct {
	ID        uuid.UUID
	Name      string
	Slug      string
	Settings  json.RawMessage
	CreatedAt time.Time
}

type Project struct {
	ID                 uuid.UUID
	OrgID              uuid.UUID
	Name               string
	Slug               string
	RepoUrl            string
	Provider           string
	AuthMode           string
	Branches           []string
	Settings           json.RawMessage
	ScheduleCron       *string
	CreatedAt          time.Time
	UpdatedAt          time.Time
	ArchivedAt         *time.Time
	WebhookSecretEnc   []byte
	WebhookSecretNonce []byte
	FailOn             string
	ScanHistory        bool
}

type Report struct {
	ID        uuid.UUID
	OrgID     uuid.UUID
	ScanID    uuid.UUID
	Format    string
	ContentGz []byte
	CreatedAt time.Time
}

type Scan struct {
	ID            uuid.UUID
	OrgID         uuid.UUID
	ProjectID     uuid.UUID
	Trigger       string
	TriggeredBy   *uuid.UUID
	Branch        string
	CommitSha     string
	CommitMessage string
	CommitAuthor  string
	Status        string
	StatusReason  string
	Partial       bool
	Summary       json.RawMessage
	GateResult    *string
	Score         *int32
	NewIssues     int32
	FixedIssues   int32
	QueuedAt      time.Time
	StartedAt     *time.Time
	FinishedAt    *time.Time
	CleanedAt     *time.Time
}

type ScanIssue struct {
	ScanID  uuid.UUID
	IssueID uuid.UUID
	OrgID   uuid.UUID
	IsNew   bool
}

type ScanTool struct {
	ID            uuid.UUID
	OrgID         uuid.UUID
	ScanID        uuid.UUID
	ToolID        string
	Name          string
	ToolVersion   string
	Status        string
	DurationMs    int64
	FindingsCount int32
	Error         string
}

type Session struct {
	ID          uuid.UUID
	UserID      uuid.UUID
	TokenHash   []byte
	MfaPending  bool
	ActiveOrgID *uuid.UUID
	Ip          string
	UserAgent   string
	CreatedAt   time.Time
	LastSeenAt  time.Time
	ExpiresAt   time.Time
}

type SshKey struct {
	ID            uuid.UUID
	OrgID         uuid.UUID
	ProjectID     uuid.UUID
	PublicKey     string
	Fingerprint   string
	PrivateKeyEnc []byte
	Nonce         []byte
	Status        string
	CreatedAt     time.Time
	RetiredAt     *time.Time
}

type User struct {
	ID            uuid.UUID
	Email         string
	Name          string
	PasswordHash  string
	TotpSecretEnc []byte
	TotpNonce     []byte
	TotpEnabled   bool
	IsSuperadmin  bool
	LockedUntil   *time.Time
	FailedLogins  int32
	CreatedAt     time.Time
	LastLoginAt   *time.Time
	Locale        string
}

type WebhookDelivery struct {
	ID         uuid.UUID
	OrgID      uuid.UUID
	ProjectID  uuid.UUID
	Provider   string
	DeliveryID string
	Event      string
	Result     string
	ReceivedAt time.Time
}